rename utils folder & fix docs for localgrapl keys
