+++ src/js/engagement_view/src/components/engagementView/sidebar/tableStructure/toggleLensTable.tsx
@@ -14,7 +14,6 @@ import { getLenses } from "services/graphQLRequests/getLenses";
 import { ToggleLensTableProps, ToggleLensTableState } from "types/CustomTypes";
 
 import { useStyles } from "../styles";
-
 const defaultToggleLensTableState = (): ToggleLensTableState => {
 	return {
 		toggled: true,
